+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Improve proxy setup.
+- Detect proxy config.
+
 ### Fixed
 - Location of .env file.
 - Removed database selection and install MariaDB if no database engine installed.